requirements.txt
numpy>=1.21.5
opencv_python>=4.5.5.62
opencv_python_headless>=4.5.5.62
Pillow>=8.4.0
PyYAML>=6.0
scikit_image==0.19.1
torch>=1.10.1
torchvision>=0.11.2
tqdm>=4.61.2
